requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
lxml==4.6.3
python-slugify==5.0.2
requests==2.25.1
Shapely==1.7.1
simplekml==1.3.5
text-unidecode==1.3
urllib3==1.26.6
